% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates_remove_seas_forecast.R
\name{covariates_remove_seas_forecast}
\alias{covariates_remove_seas_forecast}
\title{Remove Sazonalidade e Faz Previsão}
\usage{
covariates_remove_seas_forecast(
  dfts,
  start_year = 2011,
  start_period = 1,
  end_year = 2028,
  end_period = 12
)
}
\arguments{
\item{dfts}{Um objeto ts contendo covariáveis.}

\item{start_year}{Ano inicial da série.}

\item{start_period}{Período inicial (mês, se frequência mensal).}

\item{end_year}{Ano final da previsão.}

\item{end_period}{Período final da previsão.}
}
\value{
Um objeto ts contendo a série ajustada e as previsões.
}
\description{
Previsão das covariáveis
}
\details{
Esta função realiza previsões das covariáveis.
}
\examples{
covariates_remove_seas_forecast(dfts = final_data_ts)
}
