% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_gdp.R
\name{f_arima}
\alias{f_arima}
\title{Previsão do PIB}
\usage{
f_arima(y, h, seed)
}
\arguments{
\item{seed}{Número da semente para reprodutibilidade}

\item{variavel_forecasting}{Variável de interesse para previsão}

\item{horizon}{Horizonte para validação cruzada}

\item{start_date_series}{Ano-mês do início da série}

\item{start_date_series_diff}{Ano-mês do início da série}

\item{end_date_series}{Ano-mês do fim da série}

\item{start_date_forecast}{Ano-mês para para início da previsão fora da amostra}

\item{models}{Vetor com os modelos de previsão a serem considerados}

\item{regularization}{Booleano para usar regularização via LASSO}
}
\value{
Lista contendo previsões acumuladas e índices
}
\description{
Esta função realiza previsões do PIB utilizando diferentes modelos estatísticos.
}
\examples{
forecast_pib("y_pib_adjusted.servicos_ajustado", horizon = 2, start_year = 2024, start_month = 12)
}
