% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_gdp.R
\name{forecast_gdp}
\alias{forecast_gdp}
\title{Esta função realiza previsões do PIB/Inflação utilizando diversos modelos estatísticos e de aprendizado de máquina.}
\usage{
forecast_gdp(
  variavel_forecasting = "ajustado_servicos",
  start_series_year = 2011,
  start_series_period = 1,
  end_series_year = 2024,
  end_series_period = 11,
  start_forecast_year = 2024,
  start_forecast_period = 12,
  end_forecast_year = 2028,
  end_forecast_period = 12,
  horizon = 2,
  InPer = 162,
  models = c("sarima", "ets", "bats", "tbats", "prophet", "elm", "wavelet", "lasso",
    "sarimax", "gbm", "xgboost", "rf", "pcr", "mars", "hybrid"),
  seed = 1117527352,
  cross_validation = FALSE,
  regularization = TRUE,
  inflation_forecast = FALSE,
  covariates_ts = covariates_remove_seas_forecast_ts,
  responses_ts = responses_remove_seas_forecast_ts,
  names_covariates = c("crise_s", "abcr_leves_ajustado", "pms_servicos_a", "pmc_a",
    "pmc_varejo_amp_a", "transferencias_correntes", "cofins", "iof", "emprego.formal",
    "emprego.agro", "emprego.ind_ext", "emprego.formal_go", "ibcr.go_a", "ibc.br_a",
    "icc", "icea", "ief", "ics_a", "ipca_acum_12_br", "ipca", "ipca_serv", "ipca_ex1",
    "ipca_ma_sem.suavizacao", "ipca_ma_suavizado", "ipca_ex0", "ic.br_agro_us.",
    "ic_br_metal_us.", "ic_br_energia_us.", "ie_a", "selic", "selic_la12", "icms",
    "energia_consumo", "serv_a")
)
}
\arguments{
\item{variavel_forecasting}{Nome da variável de interesse para previsão.}

\item{start_series_year}{Ano de início da série temporal.}

\item{start_series_period}{Mês de início da série temporal.}

\item{end_series_year}{Ano final da série temporal.}

\item{end_series_period}{Mês final da série temporal.}

\item{start_forecast_year}{Ano de início da previsão.}

\item{start_forecast_period}{Mês de início da previsão.}

\item{end_forecast_year}{Ano final da previsão.}

\item{end_forecast_period}{Mês final da previsão.}

\item{horizon}{Horizonte de previsão para validação cruzada.}

\item{InPer}{Número de períodos iniciais utilizados para validação cruzada.}

\item{models}{Modelos a serem considerados para previsão.}

\item{seed}{Semente para reprodutibilidade dos resultados.}

\item{cross_validation}{Indica se a validação cruzada será realizada (TRUE/FALSE).}

\item{regularization}{Indica se a seleção de variáveis será feita via LASSO (TRUE/FALSE).}

\item{inflation_forecast}{Indica se a previsão é para inflação (TRUE/FALSE).}

\item{covariates_ts}{Conjunto de covariáveis para previsão.}

\item{responses_ts}{Conjunto de séries de resposta.}

\item{names_covariates}{Vetor de nomes das covariáveis utilizadas.}
}
\value{
Lista contendo previsões de crescimento e índices, além das variáveis selecionadas pelo LASSO.
}
\description{
Esta função realiza previsões do PIB/Inflação utilizando diversos modelos estatísticos e de aprendizado de máquina.
}
\examples{
forecast_gdp("ajustado_servicos")
}
