% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses_remove_seas_forecast.R
\name{responses_remove_seas_forecast}
\alias{responses_remove_seas_forecast}
\title{Remove Sazonalidade das Respostas}
\usage{
responses_remove_seas_forecast(
  dfts,
  start_year = 2011,
  start_period = 1,
  end_year = 2028,
  end_period = 12
)
}
\arguments{
\item{dfts}{Um objeto ts contendo respostas.}

\item{start_year}{Ano inicial da série.}

\item{start_period}{Período inicial (mês, se frequência mensal).}

\item{end_year}{Ano final da previsão.}

\item{end_period}{Período final da previsão.}
}
\value{
Um objeto ts contendo a série ajustada e as previsões.
}
\description{
Gerar séries sem sazonalidade
}
\details{
Esta função gera séries sem sazonalidade.
}
\examples{
responses_remove_seas_forecast(dfts = final_data_ts)
}
