% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_gdp.R
\name{forecast_gdp}
\alias{forecast_gdp}
\title{Previsão do PIB}
\usage{
forecast_gdp(
  variavel_forecasting = "y_pib_adjusted.servicos_ajustado",
  horizon = 2,
  start_date_series = c(2011, 1),
  start_date_series_diff = c(2011, 2),
  end_date_series = c(2024, 11),
  start_date_forecast = c(2024, 12),
  Start_Hor = "2024-12-01",
  Hor = 49,
  InPer = 162,
  models = c("arima", "ets", "bats", "tbats", "elm", "wavelet", "lasso", "sarimax",
    "gbm", "xgboost", "rf", "pcr", "mars", "hybrid"),
  seed = 1117527352,
  regularization = TRUE,
  pib_goias_mensal_agro_ind_serv = pib_goias_mensal_agro_ind_serv,
  covariates_ts = covariates_ts,
  y_non_seasonal_ts = y_non_seasonal_ts,
  y_seasonal_adjusted_ts = y_seasonal_adjusted_ts,
  names_covariates = c("crise_s", "abcr_leves_ajustado", "pms_servicos_a", "pmc_a",
    "pmc_varejo_amp_a", "transferencias_correntes", "cofins", "iof", "emprego.formal",
    "emprego.agro", "emprego.ind_ext", "emprego.formal_go", "ibcr.go_a", "ibc.br_a",
    "icc", "icea", "ief", "ics_a", "ipca_acum_12_br", "ipca", "ipca_serv", "ipca_ex1",
    "ipca_ma_sem.suavizacao", "ipca_ma_suavizado", "ipca_ex0", "ic.br_agro_us.",
    "ic_br_metal_us.", "ic_br_energia_us.", "ie_a", "selic", "selic_la12", "icms",
    "energia_consumo", "serv_a")
)
}
\arguments{
\item{variavel_forecasting}{Variável de interesse para previsão}

\item{horizon}{Horizonte para validação cruzada}

\item{start_date_series}{Ano-mês do início da série}

\item{start_date_series_diff}{Ano-mês do início da série}

\item{end_date_series}{Ano-mês do fim da série}

\item{start_date_forecast}{Ano-mês para para início da previsão fora da amostra}

\item{models}{Vetor com os modelos de previsão a serem considerados}

\item{seed}{Número da semente para reprodutibilidade}

\item{regularization}{Booleano para usar regularização via LASSO}
}
\value{
Lista contendo previsões acumuladas e índices
}
\description{
Esta função realiza previsões do PIB utilizando diferentes modelos estatísticos.
}
\examples{
forecast_gdp("y_pib_adjusted.servicos_ajustado", horizon = 2)
}
